backend/services/import-service/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
jinja2==3.1.2
python-multipart==0.0.9
boto3==1.34.79
pandas==2.2.2
sqlalchemy[asyncio]==2.0.29
python-jose==3.3.0
pydantic-settings>=2.9.1,<3.0
httpx==0.24.1
uvicorn
python-multipart
gunicorn
celery
jose[jwt]
psycopg[binary]==3.1.18
asyncpg>=0.27.0
python-jose==3.3.0
psycopg2
psycopg2-binary>=2.9
pg8000
jinja2